expyter/requirements.txt
atlassian-python-api==3.8.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.12
idna==2.10
oauthlib==3.1.0
prometheus-client==0.9.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
wrapt==1.12.1
